DRYed up some of the commands.

+++ jakelib/deploy.jake
@@ -20,102 +20,89 @@ namespace('deploy', function () {
   const apps = {
     consoleapi: {
       cmds: [
-        `cd ${serviceToPath('consoleapi')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         'docker build -t {{cluster_name}}-api -f Dockerfile-api . --no-cache',
         'docker tag {{cluster_name}}-api:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-api:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-api:latest',
-        'docker image prune -a -f'
       ]
     },
     'web-api': {
       cmds: [
-        `cd ${serviceToPath('web-api')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         'docker build -t {{cluster_name}}-api -f Dockerfile-api . --no-cache',
         'docker tag {{cluster_name}}-api:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-api:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-api:latest',
-        'docker image prune -a -f'
       ]
     },
     worker: {
       cmds: [
-        `cd ${serviceToPath('worker')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         'docker build -t {{cluster_name}}-{{app_name}} -f Dockerfile-worker . --no-cache',
         'docker tag {{cluster_name}}-{{app_name}}:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
-        'docker image prune -a -f'
       ]
     },
     console: {
       cmds: [
-        `cd ${serviceToPath('console')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         './node_modules/.bin/gulp docker:build --gulpfile ./gulpfile.babel.js --build={{cluster_name}}',
         'docker build -t {{cluster_name}}-{{app_name}} . --no-cache',
         'docker tag {{cluster_name}}-{{app_name}}:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
-        'docker image prune -a -f'
       ]
     },
     consumer: {
       cmds: [
-        `cd ${serviceToPath('consumer')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         './node_modules/.bin/gulp staging-gc-build',
         'docker build -t {{cluster_name}}-{{app_name}} . --no-cache',
         'docker tag {{cluster_name}}-{{app_name}}:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
-        'docker image prune -a -f'
       ]
     },
     'shipping-api': {
       cmds: [
-        `cd ${serviceToPath('shipping-api')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         'docker build -t {{cluster_name}}-{{app_name}} -f docker/api . --no-cache',
         'docker tag {{cluster_name}}-{{app_name}}:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
-        'docker image prune -a -f'
       ]
     },
     'shipping-worker': {
       cmds: [
-        `cd ${serviceToPath('shipping-worker')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         'docker build -t {{cluster_name}}-{{app_name}} -f docker/worker . --no-cache',
         'docker tag {{cluster_name}}-{{app_name}}:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
-        'docker image prune -a -f'
       ]
     },
     'inventory-worker': {
       cmds: [
-        `cd ${serviceToPath('inventory-worker')}`,
-        'eval $(aws ecr get-login --no-include-email --region us-west-2)',
         'docker build -t {{cluster_name}}-{{app_name}} -f docker/worker . --no-cache',
         'docker tag {{cluster_name}}-{{app_name}}:latest 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
         'docker push 052248958630.dkr.ecr.us-west-2.amazonaws.com/{{cluster_name}}-{{app_name}}:latest',
-        'docker image prune -a -f'
       ]
     }
   }
 
   desc('Deploy application to ECS. | [cluster_name,app_name]');
-	task('app', ['aws:loadCredentials'], { async: false }, function(cluster_name,app_name) {
-    let vars = {cluster_name, app_name};
-    let cmd = replacer(apps[app_name].cmds.join(' && '), vars);
+	task('app', ['aws:loadCredentials'], { async: false }, function(cluster_name, app_name) {
+    const path = serviceToPath(app_name);
+    if (!path) {
+      console.error(`Unknown app/service: ${app_name}.`);
+      return;
+    }
+
+    const cmds = [
+      `cd ${path}`,
+      'eval $(aws ecr get-login --no-include-email --region us-west-2)',
+      ...apps[app_name].cmds,
+      'docker image prune -a -f',
+    ];
+    const vars = {cluster_name, app_name};
+    const cmd = replacer(cmds.join(' && '), vars);
     jake.exec(cmd, { printStdout: true }, function(){
       jake.Task['ecs:restart'].execute(cluster_name, `${cluster_name}-${app_name}`);
       jake.Task['slack:deployment'].execute(cluster_name, app_name);
       complete();
     });
-
   });
 
   desc('Deploy application to ECS. | [cluster_name]');
-	task('all', ['aws:loadCredentials'], { async: false }, function(cluster_name,app_name) {
+	task('all', ['aws:loadCredentials'], { async: false }, function(cluster_name) {
     Object.keys(apps).forEach(k => {
       jake.exec(cmd, { printStdout: true }, function(){
         jake.Task['deploy:app'].execute(cluster_name, k);
